requirements.txt
functions-framework==3.*
google-cloud-translate
google-cloud-storage
streamlit
requests
streamlit-google-auth
google-cloud-secret-manager